+++ ChangeLog.md
@@ -3,6 +3,8 @@ Handlebars for XP web frontends change log
 
 ## ?.?.? / ????-??-??
 
+* Add `contains` helper - @thekid
+
 ## 0.1.0 / 2021-02-13
 
 * First public release - @thekid